feat: add table-services-products component for billing view
