+++ src/views/Billing/InvoiceDetailsView.vue
@@ -5,72 +5,7 @@
     </template>
     <template #content>
       <div class="w-full flex flex-col-reverse sm:flex-row gap-6">
-        <DataTable
-          v-model:expandedRows="expandedRows"
-          :value="data"
-          rowGroupMode="subheader"
-          groupRowsBy="region"
-          class="w-full sm:w-1/2 h-fit"
-          :pt="{
-            thead: { class: 'hidden' },
-            rowexpansioncell: 'p-0'
-          }"
-        >
-          <template #header>
-            <span class="text-base font-medium">Service and Products Charges</span>
-          </template>
-          <Column
-            expander
-            class="w-0"
-            :pt="{
-              rowToggler: { class: 'border-none' }
-            }"
-          />
-          <Column
-            field="service"
-            header=""
-            class="font-medium !text-base"
-          ></Column>
-          <Column
-            field="value"
-            header=""
-            class="font-medium !text-sm"
-            bodyStyle="text-align:right"
-          ></Column>
-          <template #expansion="slotProps">
-            <div class="p-0 m-0">
-              <DataTable
-                v-for="(expansionColuns, index) in slotProps.data.descriptions"
-                :key="index"
-                :value="expansionColuns.data"
-                class="border-none -mt-0.5 !text-sm text-color-secondary rounded-none"
-              >
-                <Column :pt="{ root: { class: 'pl-12' } }"></Column>
-                <Column
-                  field="country"
-                  :header="expansionColuns.service"
-                  class="w-full pt-4 pb-4"
-                  :pt="{
-                    bodyCell: { class: 'text-color-secondary text-sm' }
-                  }"
-                ></Column>
-                <Column
-                  field="quantity"
-                  :header="expansionColuns.quantity"
-                  :pt="{
-                    root: { class: 'pr-8' }
-                  }"
-                  bodyStyle="text-align:right"
-                ></Column>
-                <Column
-                  field="price"
-                  :header="expansionColuns.price"
-                  bodyStyle="text-align:right"
-                ></Column>
-              </DataTable>
-            </div>
-          </template>
-        </DataTable>
+        <TableServicesProducts :listProduct="listServiceProducts" />
         <div class="w-full sm:w-1/2 flex flex-col h-max border surface-border rounded-md">
           <div class="p-3 md:p-6 flex flex-col gap-4">
             <div class="flex justify-between">
@@ -165,8 +100,7 @@
   import PageHeadingBlock from '@/templates/page-heading-block'
   import PrimeButton from 'primevue/button'
   import cardFlagBlock from '@templates/card-flag-block'
-  import DataTable from 'primevue/datatable'
-  import Column from 'primevue/column'
+  import TableServicesProducts from './components/table-services-products'
 
   const props = defineProps({
     loadInvoiceDataService: {
@@ -192,8 +126,7 @@
 
   const invoiceData = ref({})
   const cardDefault = ref({})
-  const data = ref([])
-  const expandedRows = ref([])
+  const listServiceProducts = ref([])
 
   onMounted(() => {
     listServiceAndProductsChanges()
@@ -211,9 +144,11 @@
 
   const listServiceAndProductsChanges = async () => {
     try {
-      data.value = await props.listServiceAndProductsChangesService(route.params.billId)
+      listServiceProducts.value = await props.listServiceAndProductsChangesService(
+        route.params.billId
+      )
     } catch (error) {
-      data.value = []
+      listServiceProducts.value = []
     }
   }
